Vendor Gold Consumption Update
Detected a possible issue with the consumption of Gold using the new method:

While iterating items with the intention to delete them from their container, the iteration must be performed in a reverse-loop.
Iterating using a foreach (enumerator) in this context would cause a CollectionModifiedException.
Iterating using an ascending for-loop means that if the collection is modified (an item is removed), then the collection's count changes and the next item that would expect to be iterated is skipped over, having taken the previous item's index.
Reverse iterations can also fail under the latter condition, but only if multiple items are removed from the collection at once - this can be easily compensated for by checking the current iteration index against the collection's count and abstaining from accessing the collection until the index falls back within range (less than the collection count).

The other issue found here is that the ConsumeGold method was not used in all relevant contexts - if the total cost of the purchase is above Int32.MaxValue, then the usual method of consuming gold is used.

To fix these issues, I have re-written the ConsumeGold method and implemented a custom enumerator method in order to find the gold recursively, no matter how much the total cost is. Doing this made it possible to reduce the amount of code used to actually check and consume gold from different containers.

Disclaimer: My edits are untested, please verify if they perform as described above.