Add files via upload
Removed the outdated and unused ExpansionNeeded.

Since we do not have expansions people buy, or even an account property to track which expansions people have access too, or a way to enforce it, it was simply showing not needed information. Since items are enclosed in Era checks if you do not have the "Expansion" on your server you can not see it anyways.